refactory delivery-active-guard with CollectionGuard
