+++ libs/utils/src/index.ts
@@ -9,6 +9,7 @@ export * from './lib/form/validators/validators';
 export * from './lib/form/matchers';
 export * from './lib/pipes/translate-slug.module';
 export * from './lib/intercom.module';
+export * from './lib/date-helper';
 
 export { OrganizationAlgoliaResult } from './lib/algolia';
 export * from './lib/apps';
