+++ libs/utils/src/lib/date-helper.ts
@@ -1,0 +1,13 @@
+export function timestampObjectsToDate(docs: any[]) {
+  if (!docs) {
+    return [];
+  }
+
+  return docs.map(doc => {
+    if (doc.date) {
+      return { ...doc, date: doc.date.toDate() };
+    } else {
+      return doc;
+    }
+  });
+}
